Use libkdumpfile to read VMCOREINFO
If VMCOREINFO is found in the dump file itself, then libkdumpfile
has already parsed and stored it in attributes at initialization.
Let's use these attributes in this case.

However, makedumpfile has traditionally also allowed reading
VMCOREINFO data from external files. This is not directly supported
by libkdumpfile, but the VMCOREINFO raw attribute can be set to a
new value, which gets reparsed automatically.

Note that the values from the dump file should take precedence over
those found in a file, but the external VMCOREINFO may contain
additional entries, so makedumpfile is not simply chosing one of
them but rather merges both, giving the embedded VMCOREINFO a higher
priority. This can be achieved by prepending the content of the
external file to the existing attribute value, because individual
VMCOREINFO lines are parsed sequentially.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>